Fix counting logic
